scripts/requirements.txt
requests==2.33.1
beautifulsoup4==4.14.3
lxml==6.1.0
supabase==2.9.1
python-dotenv==1.1.0
numpy==2.2.5
pandas==2.2.3
lightgbm==4.6.0
scikit-learn==1.6.1
